+++ src/components/Statistics/LineChartComponent.tsx
